backend/pyproject.toml
[tool.poetry]
name = "finvest-backend"
version = "0.1.0"
description = "FinVest Platform Backend API"
authors = []
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = {version = "^0.34.0", extras = ["standard"]}
pydantic = {extras = ["email"], version = "^2.12.5"}
pydantic-settings = "^2.7.0"
sqlalchemy = "^2.0.0"
alembic = "^1.14.0"
asyncpg = "^0.30.0"
redis = "^5.2.0"
httpx = "^0.28.0"
python-jose = {version = "^3.3.0", extras = ["cryptography"]}
greenlet = "^3.3.2"
prometheus-fastapi-instrumentator = "^7.1.0"
sentry-sdk = {extras = ["fastapi"], version = "^2.54.0"}
slowapi = "^0.1.9"
python-multipart = "^0.0.22"
bcrypt = "^5.0.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.25.0"
httpx = "^0.28.0"
ruff = "^0.9.0"
mypy = "^1.14.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
target-version = "py311"
line-length = 120

[tool.mypy]
python_version = "3.11"
strict = true

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
